[IMP] stock_move_source_relocate: Refactor Code - using float_compare instead of float_zero - common unittest class
